Create working librispeech tfds config, parallelizing example creation using beam.
Simplifies the splits by creating one for each of the 7 splits in the official release (dev-clean, test-clean, dev-other,  test-other, train-clean-100, train-clean-360, train-other-500).  The result is more flexible and consistent with the way the dataset is (AFAIK) used in the literature, where it is most common to train on the union of all 3 training sets (or occasionally only on train-clean-100) and evaluate separately on each of the clean/other dev/test sets.

This means that there are no single "train", "test", or "validation" splits, but it makes it easy for users to compose different subsets as desired using the tfds Split API.

PiperOrigin-RevId: 295648622